Move plugin_contexts under api.agent_based
That's where other supporting libraries for the check api reside
(`value_store` for examples).  Moreover, we want to completely
remove `cmk.base.agent_based` so let's not re-add code there.

CMK-12298
