+++ cmk/base/agent_based/discovery/preview.py
@@ -44,7 +44,7 @@ from cmk.checkengine.sectionparserutils import check_parsing_errors
 from cmk.checkengine.summarize import SummarizerFunction
 
 import cmk.base.agent_based.checking as checking
-from cmk.base.agent_based import plugin_contexts
+from cmk.base.api.agent_based import plugin_contexts
 from cmk.base.api.agent_based.value_store import load_host_value_store, ValueStoreManager
 from cmk.base.config import ConfigCache
