+++ cmk/base/plugins/agent_based/utils/logwatch.py
@@ -31,7 +31,7 @@ from typing import (
 
 # from cmk.base.config import logwatch_rule will NOT work!
 import cmk.base.config  # pylint: disable=cmk-module-layer-violation
-from cmk.base.agent_based.plugin_contexts import (  # pylint: disable=cmk-module-layer-violation
+from cmk.base.api.agent_based.plugin_contexts import (  # pylint: disable=cmk-module-layer-violation
     host_name,
 )
 from cmk.base.plugins.agent_based.agent_based_api.v1 import regex, Result, State
