+++ tests/unit/checks/generictests/run.py
@@ -14,7 +14,7 @@ from cmk.utils.hostaddress import HostName
 
 from cmk.checkengine.checking import CheckPluginName
 
-from cmk.base.agent_based.plugin_contexts import current_host, current_service
+from cmk.base.api.agent_based.plugin_contexts import current_host, current_service
 
 from ..checktestlib import (
     assertCheckResultsEqual,
