+++ tests/unit/checks/test_mem_win.py
@@ -15,7 +15,7 @@ from cmk.utils.hostaddress import HostName
 
 from cmk.checkengine.checking import CheckPluginName
 
-from cmk.base.agent_based.plugin_contexts import current_host, current_service
+from cmk.base.api.agent_based.plugin_contexts import current_host, current_service
 from cmk.base.plugins.agent_based.agent_based_api.v1 import Metric, Result, State
 from cmk.base.plugins.agent_based.agent_based_api.v1.type_defs import CheckResult
