+++ tests/unit/cmk/base/api/agent_based/test_utils_check_levels_predictive.py
@@ -9,8 +9,8 @@ from cmk.utils.hostaddress import HostName
 
 from cmk.checkengine.checking import CheckPluginName
 
-from cmk.base.agent_based.plugin_contexts import current_host, current_service
 from cmk.base.api.agent_based import utils
+from cmk.base.api.agent_based.plugin_contexts import current_host, current_service
 from cmk.base.plugins.agent_based.agent_based_api.v1 import Result
