+++ tests/unit/cmk/base/plugins/agent_based/test_check_plugin_properties.py
@@ -17,7 +17,7 @@ from cmk.utils.hostaddress import HostName
 from cmk.checkengine.sectionparser import ParsedSectionName
 
 import cmk.base.api.agent_based.register as agent_based_register
-from cmk.base.agent_based.plugin_contexts import current_host
+from cmk.base.api.agent_based.plugin_contexts import current_host
 from cmk.base.api.agent_based.type_defs import SectionPlugin, SNMPSectionPlugin
